split preprocessing tests like module is split
